refactor(authentication): Update imports to use UserDTO from user/model
