+++ backend/src/main/java/com/project/G1_T3/authentication/controller/AuthenticationController.java
@@ -1,7 +1,7 @@
 package com.project.G1_T3.authentication.controller;
 
-import com.project.G1_T3.authentication.model.UserDTO;
 import com.project.G1_T3.authentication.service.JwtService;
+import com.project.G1_T3.user.model.UserDTO;
 import com.project.G1_T3.user.service.CustomUserDetailsService;
 import com.project.G1_T3.user.service.UserService;
